REF: Move benchmarks utils into benchmarks
